refactor : 요구사항 변경에 따른 테스트 시그니처 수정
